+++ backend/src/common/decorators/current-user.decorator.ts
@@ -10,3 +10,4 @@ export const CurrentUser = createParamDecorator(
     return request.user;
   }
 );
+
