+++ backend/src/common/decorators/public.decorator.ts
@@ -1,0 +1,8 @@
+import { SetMetadata } from '@nestjs/common';
+
+/**
+ * 公開エンドポイントを示すデコレータ
+ * JWT認証ガードで認証をスキップするために使用
+ */
+export const Public = () => SetMetadata('isPublic', true);
+
