+++ backend/src/domain/auth/auth.module.ts
@@ -3,13 +3,13 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { PrismaModule } from '../../database/prisma/prisma.module';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { AuthRepository } from './repository/auth.repository';
+import { UsersModule } from '../users/users.module';
+import { AuthController } from './controllers/auth.controller';
+import { AuthService } from './services/auth.service';
 import { AuthUsecase } from './usecase/auth.usecase';
 import { JwtStrategy } from './strategies/jwt.strategy';
 import { TokenBlacklistService } from './services/token-blacklist.service';
-import { JwtAuthGuard } from './guards/jwt-auth.guard';
+import { JwtAuthGuard } from '../../common/guards/jwt-auth.guard';
 
 /**
  * 認証モジュール
@@ -19,6 +19,7 @@ import { JwtAuthGuard } from './guards/jwt-auth.guard';
   imports: [
     PrismaModule,
     PassportModule,
+    UsersModule,
     // ConfigServiceを使用してJWT設定を動的に読み込む
     JwtModule.registerAsync({
       imports: [ConfigModule],
@@ -35,7 +36,6 @@ import { JwtAuthGuard } from './guards/jwt-auth.guard';
   ],
   controllers: [AuthController],
   providers: [
-    AuthRepository,
     AuthUsecase,
     AuthService,
     JwtStrategy,
