+++ backend/src/domain/users/repository/users.repository.ts
@@ -3,17 +3,32 @@ import { PrismaService } from '../../../database/prisma/prisma.service';
 import { User } from '../../../../generated/prisma/client';
 
 /**
+ * ユーザー作成データ型
+ */
+export type CreateUserData = {
+  email: string;
+  password: string;
+  name: string;
+  company: string;
+  failedLoginAttempts?: number;
+  lockedUntil?: Date | null;
+};
+
+/**
  * ユーザー更新データ型
  */
 export type UpdateUserData = {
   name?: string;
   company?: string;
   email?: string;
+  failedLoginAttempts?: number;
+  lockedUntil?: Date | null;
 };
 
 /**
  * ユーザーリポジトリ
  * データアクセス層を担当
+ * 認証とユーザー管理の両方の機能を提供
  */
 @Injectable()
 export class UsersRepository {
@@ -38,6 +53,17 @@ export class UsersRepository {
   async findByEmail(email: string): Promise<User | null> {
     return await this.prisma.user.findUnique({
       where: { email },
+    });
+  }
+
+  /**
+   * ユーザーを作成
+   * @param data ユーザー作成データ
+   * @returns 作成されたユーザー情報
+   */
+  async create(data: CreateUserData): Promise<User> {
+    return await this.prisma.user.create({
+      data,
     });
   }
