+++ backend/src/domain/users/users.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from '../../database/prisma/prisma.module';
 import { AuthModule } from '../auth/auth.module';
-import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
+import { UsersController } from './controllers/users.controller';
+import { UsersService } from './services/users.service';
 import { UsersUsecase } from './usecase/users.usecase';
 import { UsersRepository } from './repository/users.repository';
 
@@ -14,6 +14,6 @@ import { UsersRepository } from './repository/users.repository';
   imports: [PrismaModule, AuthModule],
   controllers: [UsersController],
   providers: [UsersRepository, UsersUsecase, UsersService],
-  exports: [UsersService],
+  exports: [UsersService, UsersRepository],
 })
 export class UsersModule {}
